mysite/requirements.txt
asgiref==3.2.10
coverage==5.3
-e git+https://github.com/django/django.git@01974d7f7549b2dca2a729c3c1a1ea7d4585eb3a#egg=Django
djangorestframework==3.12.1
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
PyMySQL==0.10.1
pytz==2020.1
sqlparse==0.3.1
